Add edit/destroy policy

+++ app/Http/Controllers/CourseController.php
@@ -49,7 +49,7 @@ class CourseController extends Controller
     public function store(Request $request)
     {
         $course = Course::create($request->all());
-        $course->tags=str_replace(' ', '', $request->tags);
+        $course->tags = str_replace(' ', '', $request->tags);
         if (!empty($request->file('image'))) {
             $path = Storage::putFile('public', $request->file('image'));
             $url = Storage::url($path);
@@ -81,10 +81,14 @@ class CourseController extends Controller
      */
     public function edit(Course $course)
     {
-        return view('course.edit', [
-            'course' => $course,
-            'delimiter' => ''
-        ]);
+        $user = Auth::user();
+
+        if ($user->can('edit', $course)) {
+            return view('course.edit', [
+                'course' => $course,
+                'delimiter' => ''
+            ]);
+        } else return redirect()->route('course.show', $course);
     }
 
     /**
@@ -96,15 +100,18 @@ class CourseController extends Controller
      */
     public function update(Request $request, Course $course)
     {
-        $course->update($request->except('slug'));
-        $course->tags=str_replace(' ', '', $request->tags);;
-        if (!empty($request->file('image'))) {
-            $path = Storage::putFile('public', $request->file('image'));
-            $url = Storage::url($path);
-            $course->image = $url;
-        }
-        $course->save();
-        return redirect()->route('course.show', $course);
+        $user = Auth::user();
+        if ($user->can('update', $course)) {
+            $course->update($request->except('slug'));
+            $course->tags = str_replace(' ', '', $request->tags);;
+            if (!empty($request->file('image'))) {
+                $path = Storage::putFile('public', $request->file('image'));
+                $url = Storage::url($path);
+                $course->image = $url;
+            }
+            $course->save();
+            return redirect()->route('course.show', $course);
+        } else return redirect()->route('course.show', $course);
     }
 
     /**
@@ -116,13 +123,11 @@ class CourseController extends Controller
      */
     public function destroy(Course $course)
     {
-//      dd($course);
-//      $user = Auth::user();
-//      if ($user->can('delete', $article)) {
-        $course->delete();
+        $user = Auth::user();
+        if ($user->can('destroy', $course)) {
+            $course->delete();
+            return redirect()->route('home');
+        } else return redirect()->route('course.show', $course);
 
-        return redirect()->route('home');
-//      } else return redirect()->route('article.index');
     }
-
 }
